requirements.txt
anthropic>=0.39.0
openai>=1.50.0
tweepy>=4.14.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
requests>=2.31.0
python-telegram-bot>=21.0
pytz>=2024.1
pytest>=8.0.0
